Learning Javascript and Node.js This is the initial commit of the bot rewrite. There was lots of learning going on with this and I had to figure out how to do everything from scratch. JS seems to actually be easier than writing in python. I have to write all the commands into different files, I haven't found a different way of doing it yet. This does make some aspects of the code easier to write and some aspects harder, like code organization and stuff.
There is too much to write for a change log. Most nothing is in the bot, just basic files that output messages to the console right now so I can test if things work. Will probably start to work with mognoDB next and see how I want to format that data and store it. It might be different because I can reference other files and such so we are going to see what happens.