FINALLY got the MongoDB to send to specific databases based off the guildID and then inserting into specific collections.
Started adding more files and adding TODOs so that I can start breaking things down and getting more things done.
Next thing todo is finish up all the inserting and updating in the database that needs to happen when things in the server are updated.
After that is to work on roleGroups, roleMenus, then a help command that has pages, which should be easier than it was in python.
In addition, twitch/youtube announcements will need to take place in the channel that the server decides if they want stream announcements.
Lastly after all that is done, cracking down and working on a web interface for the bot is the last thing after getting the basics done.